// src/enums/projects.enums.ts
export enum ActionProject{
    CREATE_PROJECT = "CREATE_PROJECT",
    UPDATE_PROJECT = "UPDATE_PROJECT",
    CREATE_TASK = "CREATE_TASK",
    UPDATE_TASK = "UPDATE_TASK",
    DELETE_TASK = "DELETE_TASK",
    CREATE_LIST = "CREATE_LIST",
    UPDATE_LIST = "UPDATE_LIST",
    ASSING_USER = "ASSING_USER"
}